monorepo/airflow/requirements.txt
kafka-python-ng
selenium==4.23.1
beautifulsoup4
webdriver-manager
